moved test_cf.py to a different folder. It will not run automatically in travis.
